+++ pywincffi/kernel32/io.py
@@ -11,8 +11,6 @@ from six import integer_types
 
 from pywincffi.core import dist
 from pywincffi.core.checks import Enums, input_check, error_check, NoneType
-from pywincffi.exceptions import WindowsAPIError
-from pywincffi.kernel32.handle import INVALID_HANDLE_VALUE
 
 PeekNamedPipeResult = namedtuple(
     "PeekNamedPipeResult",
@@ -314,21 +312,4 @@ def ReadFile(hFile, nNumberOfBytesToRead, lpOverlapped=None):
     error_check("ReadFile", code=code, expected=Enums.NON_ZERO)
     return ffi.string(lpBuffer)
 
-
-def CloseHandle(hObject):
-    """
-    Closes an open object handle.
-
-    .. seealso::
-
-        https://msdn.microsoft.com/en-us/library/ms724211
-
-    :param handle hObject:
-        The handle object to close.
-    """
-    input_check("hObject", hObject, Enums.HANDLE)
-    _, library = dist.load()
-
-    code = library.CloseHandle(hObject)
-    error_check("CloseHandle", code=code, expected=Enums.NON_ZERO)
